Deleting items from custom todolist

+++ 21. Putting Everything Together/Todolist-V2/app.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const mongoose = require('mongoose');
+const _ = require('lodash');
 
 const app = express();
 
@@ -57,7 +58,7 @@ app.get('/', (req, res) =>{
 });
 
 app.get('/:customListName', (req, res) =>{
-    const customListName = req.params.customListName;
+    const customListName = _.capitalize(req.params.customListName);
 
     List.findOne({name: customListName}, (err, foundList) =>{
         if(!err){
@@ -99,15 +100,24 @@ app.post('/', (req, res) =>{
 
 app.post('/delete', (req, res) =>{
     const checkedItemId = req.body.checkbox;
-    console.log(checkedItemId);
-    Item.findByIdAndRemove(checkedItemId, (err) =>{
-        if(!err){
-            console.log('Successfully deleted');
-            res.redirect('/');
-        } else{
-            console.log(err);
-        }
-    });
+    const listName = req.body.listName;
+
+    if(listName === 'Today'){
+        Item.findByIdAndRemove(checkedItemId, (err) =>{
+            if(!err){
+                console.log('Successfully deleted');
+                res.redirect('/');
+            } else{
+                console.log(err);
+            }
+        });
+    } else{
+        List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}}, (err, foundList) =>{
+            if(!err){
+                res.redirect('/' + listName);
+            }
+        });
+    }
 });
 
 app.listen(3000, () =>{
